+++ themes/grav/app/forms/fields/filepicker.js
@@ -38,6 +38,8 @@ export default class FilePickerField {
         let folder = '';
         let thumbs = {};
 
+        let onDemand = field.closest('[data-ondemand]').length > 0;
+
         if (!field.length || field.get(0).selectize) { return; }
 
         let getData = function getData(field, callback, mode = 'all') {
@@ -118,7 +120,9 @@ export default class FilePickerField {
             },
 
             onInitialize: function() {
-                this.load((callback) => getData(field, (data) => callback(data), 'selected'));
+                if (!onDemand) {
+                    this.load((callback) => getData(field, (data) => callback(data), 'selected'));
+                }
             },
 
             onLoad: function(/* data */) {